FGJ-12: Close resources on gateway close
Also changed OrderedBlockEventSource so that it can be constructed
with a specific start block number when used by ReplayListenerSession.
This is intended to prevent the OrderedBlockEventSource from being
initialized to an undesired start block by stray events arriving
from the low-level SDK during a replay session, and preventing the
correct blocks from being replayed.

Change-Id: I5376b572ea883331f7d12bdc84bbbf92651c53ac
Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>