Maven Dependency for Java1.9+
Beginning with Java 1.9 the jaxb-api is not included in the JDK by default anymore. See this post: http://openjdk.java.net/jeps/320
Thus, it has to be included in pom.xml.
The project  explicitly either explicitly mention that the SDK is for Java version 1.8 or it should include this dependency in the pom.xml.