FGJ-94: Only deliver contract events from committed transactions
Contract listeners are intended to allow business processes to be invoked in response to chaincode events emitted by transactions that have successfully committed and updated the ledger state. Chaincode events from transactions that are not successfully committed should not be delivered to contract event listeners as this may cause business processes to be invoked incorrectly.

Client applications are still able to interrogate chaincode events from transactions that failed to commit using a block listener and iterating over all the contained transaction events, which includes both committed and uncommitted transactions, and their associated chaincode events.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>