Fix Javadoc build for Java 17
Even if the path and compiler target point to a different Java version, the Java version associated with the
Maven CLI can cause Javadodc build failures in systems with Java 17 installed. Set an upper bound (in addition
to the lower bound) for the build profile that enables the `--no-module-directories` Javadoc option.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>