Use correct CryptoSuite and User context for event replay
Use the exact CryptoSuite and User context from the original HFClient when creating a new HFClient for event replay, instead of assuming that the CryptoSuite and User context are ones created using an X509IdentityProvider with the Gateway client identity.

Also updates dependency versions and fixed a compile warning in ContractException that could appear with Java 17.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>