Simplify pull request checks
Add a job to the end of the pull_request workflow that depends on the build jobs, and can be used as a single job to check pull request success. This avoids the need to change the branch protection PR checks when changes are made to the build process.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>
(cherry picked from commit 9e0fd3f1a82da860a2e8053b96e22aec358316b1)